AppWebBA/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
cx-Oracle==8.3.0
Django==4.1.8
django-extensions==3.2.1
djangorestframework==3.14.0
idna==3.4
marshmallow==3.17.0
mssql-django==1.2
packaging==23.1
Pillow==9.5.0
pyodbc==4.0.39
pytz==2023.3
requests==2.30.0
sqlparse==0.4.3
transbank-sdk==4.0.0
tzdata==2023.3
urllib3==2.0.2
